mcp-llm-server/requirements.txt
# Core MCP dependencies
mcp>=1.10.0

# HTTP and async support
aiohttp>=3.9.0
httpx>=0.25.0
asyncio-mqtt>=0.11.0

# Authentication
authlib>=1.3.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# LLM clients
openai>=1.35.0
anthropic>=0.28.0
google-generativeai>=0.7.0

# Configuration and environment
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Logging and monitoring
structlog>=23.1.0
rich>=13.7.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.0